backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# AI / LLM
openai==1.51.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.3
psycopg2-binary==2.9.9

# Vector DB
faiss-cpu==1.8.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-auth==2.29.0

# File Processing
python-multipart==0.0.12
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
Pillow==10.4.0
pytesseract==0.3.13

# Web Search
httpx==0.27.2
beautifulsoup4==4.12.3

# Rate Limiting
slowapi==0.1.9

# Streaming
sse-starlette==2.1.3

# Utilities
tiktoken==0.7.0
numpy==1.26.4
